+++ AutoShop.csproj
@@ -21,4 +21,9 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Pages\Partes\Edit\" />
+    <Folder Include="Pages\Partes\Delete\" />
+  </ItemGroup>
+
 </Project>
