+++ Pages/Partes/Create/Create.cshtml
@@ -1,0 +1,4 @@
+﻿@page
+@model AutoShop.Pages.Partes.Create.CreateModel
+@{
+}
